+++ package.json
@@ -15,7 +15,7 @@
   "homepage": "https://github.com/wednesday-solutions/reporting-dashboard#readme",
   "scripts": {
     "start:local": "nodemon",
-    "local": "kill-port 9000 && export ENVIRONMENT=local && yarn run build:local && yarn run start",
+    "local": "kill-port 9000 && export ENVIRONMENT=local && yarn build:local && yarn start",
     "build:local": "export ENVIRONMENT_NAME=local && rm -rf dist && webpack --mode development --config  webpack.dev.config.js",
     "build:development": "export ENVIRONMENT_NAME=development && rm -rf dist && webpack --mode development --config  webpack.dev.config.js",
     "build:qa": "export ENVIRONMENT_NAME=qa && rm -rf dist && webpack --mode development --config  webpack.dev.config.js",
@@ -23,10 +23,10 @@
     "start": "node ./dist/main.js",
     "clean": "rimraf dist",
     "format": "prettier-standard './**/**/*.js'",
-    "lint": "yarn run lint:js",
+    "lint": "yarn lint:js",
     "lint:eslint": "eslint --ignore-path .gitignore --ignore-pattern server/bin",
     "lint:eslint:fix": "eslint --ignore-path .gitignore --ignore-pattern server/bin --fix",
-    "lint:js": "yarn run lint:eslint -- . ",
+    "lint:js": "yarn lint:eslint -- . ",
     "lint:staged": "lint-staged",
     "prettify": "prettier --write",
     "precommit": "lint:staged",
