+++ resources/v1/08_users.sql
@@ -9,3 +9,5 @@ CREATE TABLE users (
     deleted_at timestamp WITH time zone
 );
 
+CREATE INDEX user_email ON users USING btree (email);
+
