+++ server/gql/mutations.js
@@ -8,7 +8,6 @@ import { addressMutations } from '@gql/models/addresses';
 import { storeMutations } from '@gql/models/stores';
 import { storeProductMutations } from '@gql/models/storeProducts';
 import { supplierProductMutations } from '@gql/models/supplierProducts';
-import { userMutations } from '@gql/models/users';
 
 export const createResolvers = model => ({
   createResolver: (parent, args, context, resolveInfo) => model.create(args),
@@ -22,8 +21,7 @@ export const DB_TABLES = {
   store: storeMutations,
   storeProduct: storeProductMutations,
   supplier: supplierMutations,
-  supplierProduct: supplierProductMutations,
-  user: userMutations
+  supplierProduct: supplierProductMutations
 };
 
 export const addMutations = () => {
