+++ server/utils/testUtils/mockData.js
@@ -1,5 +1,7 @@
 import range from 'lodash/range';
 import faker from 'faker';
+import md5 from 'md5';
+const createdBefore = parseInt(Math.random() * 1000);
 
 export const addressesTable = range(1, 10).map((_, index) => ({
   id: (index + 1).toString(),
@@ -15,7 +17,9 @@ export const usersTable = range(1, 10).map((_, index) => ({
   id: (index + 1).toString(),
   firstName: faker.name.firstName(),
   lastName: faker.name.lastName(),
-  email: faker.internet.email()
+  email: faker.internet.email(),
+  password: md5(faker.internet.password()),
+  created_at: faker.date.recent(createdBefore)
 }));
 
 export const productsTable = range(1, 10).map((_, index) => ({
